backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
websockets
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
httpx
python-jose[cryptography]
passlib[bcrypt]
pytest
pytest-asyncio
pytest-mock
google-generativeai
ollama
